function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.11.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
sendgrid==6.9.7
